# src/main/resources/swagger-schema/task.swagger.yaml
openapi: 3.0.3
info:
  title: Task API
  description: API для управления задачами
  version: 1.0.0
servers:
  - url: http://localhost:8087/api/tasks

paths:
  /task:
    post:
      summary: "Создание новой задачи"
      description: "Можно создать новую задачу, введя данные."
      tags:
        - Задачи
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: "Задача создана"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: "Ошибка валидации данных"

  /task/{id}:
    delete:
      summary: "Удаление задачи"
      description: "Можно удалить задачу, введя ID задачи."
      tags:
        - Задачи
      parameters:
        - name: id
          in: path
          required: true
          description: "ID задачи для удаления"
          schema:
            type: integer
      responses:
        '204':
          description: "Задача удалена"
        '404':
          description: "Задача не найдена"
        '500':
          description: "Ошибка при удалении задачи"

    put:
      summary: "Обновление данных по задаче"
      description: "Можно обновить данные по задаче, введя ID задачи."
      tags:
        - Задачи
      parameters:
        - name: id
          in: path
          required: true
          description: "ID задачи для обновления"
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: "Задача обновлена"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: "Задача не найдена"
        '400':
          description: "Ошибка валидации данных"

  /tasks:
    get:
      summary: "Получение всех задач"
      tags:
        - Задачи
      responses:
        '200':
          description: "Список всех задач"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

  /user/{userId}:
    get:
      summary: "Получение задач по ID исполнителя"
      description: "Можно получить задачи по ID исполнителя."
      tags:
        - Задачи
      parameters:
        - name: userId
          in: path
          required: true
          description: "ID исполнителя"
          schema:
            type: string
            format: uuid
        - name: size
          in: query
          required: false
          description: "Размер страницы"
          schema:
            type: integer
            default: 5
        - name: page
          in: query
          required: false
          description: "Номер страницы"
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: "Список задач для исполнителя"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'

  /search:
    get:
      summary: "Получение задач по фамилии исполнителя"
      description: "Можно получить задачи по фамилии исполнителя."
      tags:
        - Задачи
      parameters:
        - name: lastName
          in: query
          required: true
          description: "Фамилия исполнителя"
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: "Номер страницы"
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          description: "Размер страницы"
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: "Список задач по фамилии исполнителя"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '404':
          description: "Задачи не найдены"
        '400':
          description: "Ошибка валидации данных"

components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          description: "Уникальный идентификатор задачи"
        title:
          type: string
          description: "Название задачи"
        description:
          type: string
          description: "Описание задачи"
        userId:
          type: string
          description: "ID пользователя, которому назначена задача"
          format: uuid
        comment:
          type: string
          description: "Комментарий к задаче"
        status:
          type: string
          description: "Статус задачи"
          enum:
            - NEW
            - IN_PROGRESS
            - COMPLETED
        priority:
          type: string
          description: "Приоритет задачи"
          enum:
            - LOW
            - MEDIUM
            - HIGH